requirements.txt
streamlit==1.53.1
sqlalchemy==2.0.25
pandas==2.2.0
plotly==5.18.0
reportlab==4.0.9
python-dateutil==2.8.2
